+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Don't retry incoming activity if fetcher encounters `404 Not Found` error.
 - Rename `check-expired-invoice` command to `reopen-invoice` and allow to reopen invoices with "forwarded" status.
 - Allow to call `reopen-invoice` command with payment address as an argument.
+- Verify that account index returned by monero-wallet-rpc matches configuration.
 
 ### Deprecated